requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.1
async-timeout==3.0.1
attrs==21.4.0
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
discord.py==1.7.3
ffmpeg==1.4
idna==3.3
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==6.0.2
pycparser==2.20
pylint==2.7.2
pymongo==3.12.1
PyNaCl==1.5.0
pyspnego==0.1.5
six==1.15.0
smbprotocol==1.5.0
toml==0.10.2
typing_extensions==4.1.1
wrapt==1.12.1
yarl==1.7.2
youtube-dl==2021.12.17
